% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_stat.R
\name{CalculateFP}
\alias{CalculateFP}
\title{Calculations for the friendship paradox}
\usage{
CalculateFP(A)
}
\arguments{
\item{A}{an adjacency \code{\link{matrix}}}
}
\value{
\code{\link{list}}. The first vector,
        \code{$degree}, indicates the type of degree calculated,
        \code{$kmean}, mean degree,
        \code{$kff}, mean number of friends of friends,
        \code{ratio}, ratio between kff and kmean,
        \code{variance}, the variance of degree distribution,
        \code{difference}, the difference between the mean degree of friends and the 
        mean degree.
}
\description{
Function to calculate friendship paradox metrics
}
\details{
This is a function that calculates the friendship paradox metrics
}
\examples{
# Loading data from....

# call function
CalculateFP(A)
                                                   
}
\references{
[1] Feld, S.L., 1991. 
    Why Your Friends Have More Friends than You Do. 
    Am. J. Sociol. 96, 1464–1477.
    
[2] Amaku, M., Cipullo, R.I., Grisi-Filho, J.H.H., Marques, F.S., Ossada, R., 2014. 
    The friendship paradox in scale-free networks. 
    Appl. Math. Sci. 8, 9. doi:10.12988/ams.2014.4288


\url{http://github.com/leb-fmvz-usp.github/epinemo}
}

