% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringCoefficient.R
\name{clusteringCoefficient}
\alias{clusteringCoefficient}
\title{Clustering Coefficient}
\usage{
clusteringCoefficient(A, directed = T)
}
\arguments{
\item{A}{an adjacency \code{\link{matrix}} with network information.}

\item{directed}{\code{\link{logical}} whether to calculate the directed or undirected clustering coefficients.}
}
\value{
If directed = TRUE, a \code{\link{data.frame}}. The first column,
        \code{$clustering.in}, is the inward clustering coefficient,
        \code{$clustering.out}, is the outward clustering coefficient.
        If directed = FALSE, a \code{\link{vector}}, with the undirected clustering coefficient.
}
\description{
A function to calculate the clustering coefficient for all nodes in a network.
}
\details{
This function calculates the inward and outward clustering coefficient in directed networks [1].
}
\examples{
# Generate an arbitrary 100 by 100 adjacency matrix with zeros and ones
# Remove loops
A <- matrix(rbinom(100 * 100, 1, 0.2), ncol = 100, nrow = 100)
diag(A) <- 0

# call contact chain function
cc <- clusteringCoefficient(A)
                                                   
}
\references{
[1] Caldarelli G (2007). Scale-Free Networks. Oxford University Press, Oxford.
}
