% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CreateUniqueIds.R
\name{CreateUniqueIds}
\alias{CreateUniqueIds}
\title{Create Unique Identifiers}
\usage{
CreateUniqueIds(data, from, to)
}
\arguments{
\item{data}{Edgelist of a network. \code{\link{data.frame}}}

\item{from}{\code{\link{character}} \code{\link{vector}}, indicate the column(s) used to characterize the source node of each link}

\item{to}{\code{\link{character}} \code{\link{vector}}, indicate the column(s) used to characterize each the target node of each link}
}
\value{
\code{\link{list}} containing 2 elements: a \code{\link{data.frame}}
         with database and network IDs for each node; and the original dataset with
         the new IDs, ranging from 1:n (columns "From" and "To")
}
\description{
Create unique identifiers in a network presented in edgelist form
}
\details{
Used to be slow, now is very fast!
}
\examples{
# New id's
new_database <- CreateUniqueId(database)
}
\references{
None.
\url{http://github.com/leb-fmvz-usp.github/epinemo}
}

