% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendshipParadox.R
\name{friendshipParadox}
\alias{friendshipParadox}
\title{Calculate parameters related to the friendship paradox}
\usage{
friendshipParadox(A)
}
\arguments{
\item{A}{an adjacency \code{\link{matrix}}}
}
\value{
\code{\link{list}}. The first vector,
        \code{$degree}, indicates the type of degree calculated,
        \code{$kmean}, mean degree,
        \code{$kff}, mean number of friends of friends,
        \code{ratio}, ratio between kff and kmean,
        \code{variance}, the variance of degree distribution,
        \code{difference}, the difference between the mean degree of friends and the 
        mean degree.
}
\description{
Function to calculate friendship paradox metrics
}
\details{
Function to calculate friendship paradox metrics
}
\examples{
# Loading data from....

# call function
CalculateFP(A)
                                                   
}
\references{
[1] Feld SL (1991). 
    Why Your Friends Have More Friends than You Do. 
    American Journal of Sociology 96, 1464–1477.
    
[2] Amaku M, Cipullo RI, Grisi-Filho JHH, Marques FS, Ossada R (2014). 
    The Friendship Paradox in Scale-Free Networks. 
    Applied Mathematical Sciences, 8 (37), 1837-1845. doi:10.12988/ams.2014.4288


\url{http://github.com/leb-fmvz-usp.github/epinemo}
}
