% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkRank.R
\name{LinkRank}
\alias{LinkRank}
\title{LinkRank Matrix}
\usage{
LinkRank(G, pr)
}
\arguments{
\item{G}{Google Matrix. Output of \code{\link{GoogleMatrix}} function}

\item{pr}{PageRank vector. Output of \code{\link{PageRank}} function}
}
\value{
LinkRank Matrix
}
\description{
Calculates the LinkRank Matrix, according to Kim et al. [1]
}
\details{
Complicated function... see the paper!
}
\examples{
# Calculates LinkRank Matrix
LinkRank(G,pr)
}
\references{
[1] Kim Y, Son SW, Jeong H (2010). 
"Finding Communities in Directed Networks." 
Physical Review E 81, 016103.
\doi{10.1103/PhysRevE.81.016103}

\url{http://github.com/leb-fmvz-usp.github/epinemo}
}
