% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkRankMod.R
\name{LinkRankMod}
\alias{LinkRankMod}
\title{LinkRank Modularity}
\usage{
LinkRankMod(L, pr, c)
}
\arguments{
\item{L}{LinkRank Matrix. Output of \code{\link{LinkRank}} function}

\item{pr}{PageRank vector. Output of \code{\link{PageRank}} function}

\item{c}{partition vector}
}
\value{
LinkRank Modularity value
}
\description{
Computes the LinkRank Modularity (\emph{Qlr}), as described by Kim et al. [1]
}
\details{
The LinkRank Modularity (\emph{Qlr}) is a modified modularity
for both directed and undirected networks, defined as [1]
\deqn{Qlr = (fraction of time spent walking within communities by a random walker)
- (expected value of this fraction).}
According to this definition, "a community is no longer a group of nodes
in which links are more densely located.
Instead, a community is a group of nodes in which a random walker
is more likely to stay" [1].
}
\examples{
# Calculates LinkRank Modularity
LinkRankMod(L,pr,c)
}
\references{
[1] Kim Y, Son SW, Jeong H (2010). 
"Finding Communities in Directed Networks." 
Physical Review E 81, 016103.
\doi{10.1103/PhysRevE.81.016103}

\url{http://github.com/leb-fmvz-usp.github/epinemo}
}
