% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PageRank.R
\name{pageRank}
\alias{pageRank}
\title{PageRank of a Network Matrix}
\usage{
pageRank(A, alpha = 0.85, epsilon = 1e-08, pr0, v)
}
\arguments{
\item{A}{network adjacency \code{\link{matrix}}}

\item{alpha}{scaling parameter in PageRank model. It must be \code{\link{numeric}} between 0 and 1 inclusive. Default = 0.85.}

\item{epsilon}{convergence tolerance. Default = 1e-8.}

\item{pr0}{starting vector at iteration 0 (a row vector). Usually set to rep(1/n,n).}

\item{v}{personalization vector. sum(v) = 1.}
}
\value{
Numeric vector with nodes' PageRank
}
\description{
Calculates the PageRank for a network matrix.
}
\details{
This function uses the power method to calculate the PageRank
of a network matrix [1].
}
\examples{
# Generate an arbitrary 100 by 100 adjacency matrix with zeros and ones
# Remove loops
A <- matrix(rbinom(100 * 100, 1, 0.2), ncol = 100, nrow = 100)
diag(A) <- 0

# Calculates PageRank
pageRank(A)
}
\references{
[1] Langville AN, Meyer CD (2006). "Google's PageRank and Beyond: The Science 
of Search Engine Rankings." Princeton University Press, Princeton.
}
