% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PageRank_H.R
\name{PageRank_H}
\alias{PageRank_H}
\title{PageRank of a Row Normalized Hyperlink Matrix}
\usage{
PageRank_H(H, alpha = 0.85, epsilon = 1e-08, pr0, v)
}
\arguments{
\item{H}{\code{\link{matrix}} Row Normalized Hyperlink Matrix. N-by-N sparse matrix. Output of \code{\link{Rnhm}} function.}

\item{alpha}{scaling parameter in PageRank model. It must be \code{\link{numeric}} between 0 and 1 inclusive. Default = 0.85.}

\item{epsilon}{convergence tolerance. Default = 1e-8.}

\item{pr0}{starting vector at iteration 0 (a row vector). Usually set to rep(1/n,n).}

\item{v}{personalization vector. sum(v) = 1.}
}
\value{
Numeric vector with nodes' PageRank
}
\description{
Calculates the PageRank for an n-by-n Markov Matrix (Row Normalized Hyperlink Matrix).
}
\details{
This function uses the power method to calculate
the PageRank of a row normalized hyperlink matrix [1].
}
\examples{
# Calculates PageRank
PageRank_H(rnhm(A))
}
\references{
[1] Langville AN, Meyer CD (2006). "Google's PageRank and Beyond: The Science 
of Search Engine Rankings." Princeton University Press, Princeton.

\url{http://github.com/leb-fmvz-usp.github/epinemo}
}
