% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRS2nets.R
\name{SIRS2nets}
\alias{SIRS2nets}
\title{Function to run the simulation of disease spread considering two networks}
\usage{
SIRS2nets(M_adj1, M_adj2, pspread1, pspread2, tSim, tI, tImin, tImax, tR, tRmin,
  tRmax, tcontrol)
}
\arguments{
\item{Two}{adjacency \code{\link{matrix}}}
}
\value{
\code{\link{list}}. The first element, a
        \code{\link{matrix}} of Infected, and the second element, a
        \code{\link{matrix}} of Recovered.
}
\description{
Function to run the simulation of disease spread considering two networks
}
\details{
Function to run the simulation of disease spread on a network
}
\examples{
# Loading data from....

# call function

                                                   
}
\references{
\url{http://github.com/leb-fmvz-usp.github/epinemo}
}

