% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringCoefficient.R
\name{clusteringCoefficient}
\alias{clusteringCoefficient}
\title{Clustering Coefficient}
\usage{
clusteringCoefficient(A, directed = T)
}
\arguments{
\item{A}{an adjancency \code{\link{matrix}} with network information.}

\item{directed}{\code{\link{logical}} wether to calculate the directed or undirected clustering coefficients.}
}
\value{
If directed = TRUE, a \code{\link{data.frame}}. The first column,
        \code{$clustering.in}, is the inward clustering coefficient,
        \code{$clustering.out}, is the outward clustering coefficient.
        If directed = FALSE, a \code{\link{vector}}, with the undirected clustering coefficient.
}
\description{
Function to calculate the clustering coefficient for all nodes in a network.
}
\details{
This function calculates the Inward and Outward clustering coefficient (formulas) in directed networks.
}
\examples{
# Loading data from....

# call contact chain function
cc <- clusteringCoefficient(A)
                                                   
}
\references{
[1] CALDARELLI, G. Scale-Free Networks. Oxford: Oxford University Press, 2007.
}
