% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendshipParadox.R
\name{friendshipParadox}
\alias{friendshipParadox}
\title{Friendship paradox metrics}
\usage{
friendshipParadox(A)
}
\arguments{
\item{A}{An adjacency \code{\link[base]{matrix}}.}
}
\value{
A list of 
\item{degree}{the type of degree calculated: indegree (\code{kin}), outdegree (\code{kout}) 
and total degree (\code{k}).}
\item{kmean}{the mean number of friends (degree).}
\item{kff}{the mean number of friends of friends (degree of friends).}
\item{ratio}{the ratio between \code{kff} and \code{kmean}.}
\item{variance}{the variance of degree distribution.}
\item{difference}{the difference between \code{kff} and \code{kmean}.}
}
\description{
Calculate parameters related to the friendship paradox
}
\details{
This function calculates the following friendship paradox metrics: mean degree (\code{kmean}),
mean degree of friends (\code{kff}), ratio between \code{kff} and \code{kmean} 
(\code{ratio}), variance of the degree distribution (\code{variance}) and difference 
between the mean degree of friends and the mean degree (\code{difference}).
}
\examples{
require(Matrix)

# Generate an arbitrary 10 by 10 adjacency matrix with zeros and ones
# Remove loops
A <- matrix(rbinom(10 * 10, 1, 0.2), ncol = 10, nrow = 10)
diag(A) <- 0

# Call function
friendshipParadox(A)
                                                   
}
\references{
[1] Feld SL (1991). 
"Why Your Friends Have More Friends than You Do." 
American Journal of Sociology, 96, 1464-1477.
    
[2] Amaku M, Cipullo RI, Grisi-Filho JHH, Marques FS, Ossada R (2014). 
"The Friendship Paradox in Scale-Free Networks." 
Applied Mathematical Sciences, 8 (37), 1837-1845. 
\doi{10.12988/ams.2014.4288}

[3] Amaku M, Grisi-Filho JHH, Negreiros RL, Dias RA, Ferreira F, Ferreira Neto JS,
Cipullo RI, Marques FS, Ossada R (2015). 
"Infectious Disease Surveillance in Animal Movement Networks: 
An Approach Based on the Friendship Paradox."
Preventive Veterinary Medicine, 121, 306-313. 
\doi{10.1016/j.prevetmed.2015.08.002}

[4] Amaku M, Grisi-Filho JHH (2015). "The Friendship Paradox as a Strategy for Scenarios 
with Incomplete Network Data." 
Physics of Life Reviews, 15, 39-40. 
\doi{10.1016/j.plrev.2015.07.006}
}
