% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoogleMatrix.R
\name{GoogleMatrix}
\alias{GoogleMatrix}
\title{This function creates the Google Matrix}
\usage{
GoogleMatrix(A, alpha = 0.85)
}
\arguments{
\item{A}{network adjacency \code{\link{matrix}}}

\item{alpha}{scaling parameter in PageRank model. It must be \code{\link{numeric}} between 0 and 1 inclusive. Default = 0.85.}
}
\value{
Google \code{\link{matrix}}.
}
\description{
A function to calculate the Google Matrix.
}
\details{
This function is very simple.
}
\examples{
# Calculates Google Matrix
Google.matrix(A)
}
\references{
[1] Page L, Brin S, Motwani R, Winograd T (1999). "The PageRank 
Citation Ranking: Bringing Order to the Web." Stanford Digital 
Library Technologies Project 1999.
\url{http://ilpubs.stanford.edu:8090/422/1/1999-66.pdf}

[2] Langville AN, Meyer CD (2006). "Google's PageRank and Beyond: The Science 
of Search Engine Rankings." Princeton University Press, Princeton.

\url{http://github.com/leb-fmvz-usp.github/epinemo}
}
