% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkRankMod.R
\name{LinkRankMod}
\alias{LinkRankMod}
\title{LinkRank Modularity}
\usage{
LinkRankMod(L, pr, c)
}
\arguments{
\item{L}{LinkRank Matrix. Output of \code{\link{LinkRank}} function}

\item{pr}{PageRank vector. Output of \code{\link{PageRank}} function}

\item{c}{partition vector}
}
\value{
LinkRank Modularity value
}
\description{
Computes the Link Rank Modularity (Qlr), as described by Kim et al. [1]
}
\details{
Complicated function... see the paper!
}
\examples{
# Calculates LinkRank Modularity
LinkRankMod(L,pr,c)
}
\references{
[1] Kim Y, Son SW, Jeong H (2010). 
"Finding Communities in Directed Networks." 
Physical Review E 81, 016103.
\doi{10.1103/PhysRevE.81.016103}

\url{http://github.com/leb-fmvz-usp.github/epinemo}
}
