% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{Makes a graph with Pareto's Law}
\usage{
pareto(A1, xlabel = "More connected nodes (\%)", ylabel = "\% of total",
  legin = "in", legout = "out")
}
\arguments{
\item{A1}{Adjacency \code{\link{matrix}}}

\item{xlabel}{\code{\link{character}} with x-axis label}

\item{ylabel}{\code{\link{character}} with y-axis label}

\item{legin}{\code{\link{character}} with legend for K_in}

\item{legin}{\code{\link{character}} with legend for K_out}
}
\value{
\code{\link{graph}} with Pareto's Law
}
\description{
Function to make a graph with Pareto's Law
}
\details{
Graph with Pareto's Law
}
\examples{

pareto(A1)

}
\references{
\url{http://github.com/leb-fmvz-usp.github/epinemo}
}

