% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnhm.R
\name{Rnhm}
\alias{Rnhm}
\title{Row-Normalized Hyperlink Matrix}
\usage{
Rnhm(A)
}
\arguments{
\item{A}{network adjacency \code{\link{matrix}}}
}
\value{
row-normalized hyperlink matrix
}
\description{
This function creates a sparse row-normalized hyperlink matrix.
}
\details{
This function creates a sparse row-normalized hyperlink matrix, 
defined in more detail in [1].
}
\examples{
# Calculates rnhm
Rnhm(A)
}
\references{
[1] Langville AN, Meyer CD (2006). "Google's PageRank and Beyond: The Science 
of Search Engine Rankings." Princeton University Press, Princeton.

\url{http://github.com/leb-fmvz-usp.github/epinemo}
}
