% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Rnhm}
\alias{Rnhm}
\title{row-normalized hyperlink matrix}
\usage{
Rnhm(A)
}
\arguments{
\item{A}{network adjacency \code{\link{matrix}}}
}
\value{
row-normalized hyperlink matrix
}
\description{
This function creates a sparse row-normalized hyperlink matrix
}
\details{
Complicated function... see the book!
}
\examples{
# Calculates rnhm
Rnhm(A)
}
\references{
Livro do PageRank

\url{http://github.com/leb-fmvz-usp.github/epinemo}
}

