% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulationSIRS.R
\name{simulationSIRS}
\alias{simulationSIRS}
\title{Makes simulations in a network}
\usage{
simulationSIRS(M_adjc, pspreadc, tSim, tI, tImin, tImax, tR, tRmin, tRmax)
}
\arguments{
\item{M_adjc}{Adjacency \code{\link{matrix}}}

\item{pspreadc}{\code{\link{numeric}}, indicates the probability of spread}

\item{tSim}{\code{\link{numeric}}, indicates the simulation time}

\item{tI}{\code{\link{vector}}, indicates the infected nodes at t = 0}

\item{tImin}{\code{\link{numeric}}, indicates the minimum duration of infection}

\item{tImax}{\code{\link{numeric}}, indicates the maximum duration of infection}

\item{tR}{\code{\link{vector}}, indicates the recovered nodes at t = 0}

\item{tRmin}{\code{\link{numeric}}, indicates the minimum duration of recovery}

\item{tRmax}{\code{\link{numeric}}, indicates the maximum duration of recovery}
}
\value{
\code{\link{list}} with two elements. The first, M_Sim_I, M_Sim_R
        \code{*$M_Sim_I}, with one column for each time step and one line for each node. Cells indicate the node's state
        (infected or not) in a given time step.
        \code{*$M_Sim_R}, with one column for each time step and one line for each node. Cells indicate the node's state
        (recovered or not) in a given time step.
}
\description{
Function to simulate disease spread in a network
}
\details{
SIRS model
}
\examples{
simulationSIRS( matriz, pspreadc = probabilidade, tSim = tsim, tI = tI, tImin = timin, tImax = timax, tR = tR, tRmin = trmin, tRmax = trmax)
}
\references{
\url{http://github.com/leb-fmvz-usp.github/epinemo}
}

