% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationSIRS.R
\name{simulationSIRS}
\alias{simulationSIRS}
\title{Function to run the simulation of disease spread on a network}
\usage{
simulationSIRS(M_adj, pspread, tSim, tI, tImin, tImax, tR, tRmin, tRmax,
  tcontrol)
}
\arguments{
\item{A}{an adjacency \code{\link{matrix}}}
}
\value{
\code{\link{list}}. The first element, a
        \code{\link{matrix}} of Infected, and the second element, a
        \code{\link{matrix}} of Recovered.
}
\description{
Function to run the simulation of disease spread on a network
}
\details{
Function to run the simulation of disease spread on a network
}
\examples{
# Loading data from....

# call function

                                                   
}
\references{
\url{http://github.com/leb-fmvz-usp.github/epinemo}
}
